Fix the issue of not submitting the form on Safari
On Safari: when an element is get with jQuery, and then removed with JS, it's still there.

That problem causes the WordPress's check in wp-admin/js/post.js fails: the check for $( '#submitdiv' ) at line 701 and the updateText() function at line 736.

To avoid the problem, we must ensure the WordPress's JS code runs AFTER the app finished rendering. Or make WordPress can't detect the elements. We do that by:

- Force the app to render with React 17, which is synchronous (not ideally).
- Empty the #poststuff div first (in the edit.js file) when the file is loaded. So no #submitdiv or #timestampdiv.
