+++ app/post-type/App.js
@@ -1,4 +1,4 @@
-import { createRoot } from "@wordpress/element";
+import { render } from "@wordpress/element";
 import { SettingsProvider } from '../SettingsContext';
 import DefaultSettings from './constants/DefaultSettings';
 import MainTabs from './MainTabs';
@@ -11,33 +11,6 @@ const container = document.getElementById( 'poststuff' );
 container.classList.add( 'mb-cpt' );
 container.id = 'mb-cpt-app';
 
-const root = createRoot( container );
-root.render( <App /> );
-
-// Remove .wp-header-end element to properly show notices.
-document.querySelector( '.wp-header-end' ).remove();
-
-const form = document.querySelector( '#post' );
-
-// Force form to validate to force users to enter required fields.
-// Use setTimeout because this attribute is dynamically added.
-setTimeout( () => {
-	form.removeAttribute( 'novalidate' );
-}, 100 );
-
-// Set post status when clicking submit buttons.
-form.addEventListener( 'submit', e => {
-	const submitButton = e.submitter;
-	const status = submitButton.dataset.status;
-	const originalStatus = document.querySelector( '#original_post_status' ).value;
-	if ( originalStatus !== status ) {
-		document.querySelector( '[name="messages"]' ).setAttribute( 'name', MbbApp.status !== 'publish' ? 'publish' : 'save' );
-	}
-	if ( originalStatus === 'auto-draft' && status === 'draft' ) {
-		document.querySelector( '[name="messages"]' ).setAttribute( 'name', 'save' );
-	}
-
-	submitButton.disabled = true;
-	submitButton.setAttribute( 'value', MbbApp.saving );
-	document.querySelector( '[name="post_status"]' ).setAttribute( 'value', status );
-} );
+// Use React 17 to make the rendering synchronous to make sure WordPress's JS (like detecting #submitdiv or .wp-header-end)
+// runs after the app is rendered.
+render( <App />, container );
